+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Increment the:
   [#2301](https://github.com/open-telemetry/opentelemetry-cpp/pull/2301)
 * [EXPORTER] Remove explicit timestamps from metric points exported by Prometheus
   [#2324](https://github.com/open-telemetry/opentelemetry-cpp/pull/2324)
+* [EXPORTER] Handle attribute key collisions caused by sanitation
+  [#2324](https://github.com/open-telemetry/opentelemetry-cpp/pull/2326)
 
 ## [1.11.0] 2023-08-21